requirements.txt
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
openai>=1.12.0
pydantic-ai>=0.1.0